refactor: restructure the directories
